refactor files, add authentication
