+++ assets/template/login.html
@@ -15,8 +15,8 @@
 
 <body class="text-center">
     <main class="form-signin">
-        <form action="/sign-in" method="post"> 
-            <img class="mb-4" src="/assets/img/logo.svg" alt="" width="64">
+        <form action="/login" method="post"> 
+            <img class="mb-4" src="/assets/img/keyboard.svg" alt="" width="64">
 
             <div class="form-floating">
                 <input type="text" class="form-control" id="username" name="username" placeholder="User Name">
@@ -27,10 +27,7 @@
                 <label for="passwd">Password</label>
             </div>
 
-            <div class="checkbox mb-3">
-                <label><input type="checkbox" name="remember" value="true"> Remember me</label>
-            </div>
-            <button class="w-100 btn btn-lg btn-primary" type="submit">Sign in</button>
+            <button class="w-100 btn btn-lg btn-primary mt-5" type="submit">Sign in</button>
             <p class="mt-5 mb-3 text-muted">WiTTY: Web-based Interactive TTY</p>
         </form>
     </main>
